requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI & LLM (new google-genai package)
google-genai>=1.0.0

# Observability 
opik>=1.0.0

# Computer Vision
opencv-python>=4.9.0.80
ultralytics>=8.1.0
mediapipe>=0.10.9

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
pillow>=10.2.0

# Deployment
mangum>=0.17.0
